refactor: useMidpoint 커스텀 훅 추가
