+++ frontend/src/hooks/index.js
@@ -1,5 +1,6 @@
 export { useConfirm } from './useConfirm';
 export { useMapView } from './useMapView';
+export { useMidpoint } from './useMidpoint';
 export { useModal } from './useModal';
 export { useParticipantAddressSearch } from './useParticipantAddressSearch';
 export { useParticipantAddressInput } from './useParticipantAddressInput';
