+++ frontend/src/pages/Midpoint/index.js
@@ -1,38 +1,14 @@
 import React, { useContext, useEffect } from 'react';
-import { useQuery } from 'react-query';
 
 import { ParticipantList } from '../../components';
 import { MapViewArea, MapView, ContentArea, ListSection, ResultSection } from './style';
 import { ParticipantContext } from '../../contexts';
-import { useMapView } from '../../hooks';
-import { API_URL } from '../../constants';
-import { httpRequest } from '../../utils';
+import { useMapView, useMidpoint } from '../../hooks';
 
 export const MidpointPage = () => {
   const { participants } = useContext(ParticipantContext);
   const { mapViewRef, showMapView, setMarker, setMarkers, setBounds } = useMapView();
-
-  const fetchMidpoint = async ({ queryKey }) => {
-    const [_, participants] = queryKey;
-    const locations = participants.map(({ x, y }) => ({ x, y }));
-    const res = await httpRequest.post(API_URL.MIDPOINT, { body: { locations } });
-
-    return await res.json();
-  };
-
-  const fetchStations = async ({ queryKey }) => {
-    const [category, midpoint] = queryKey;
-    const res = await httpRequest.get(API_URL.CATEGORY(category, midpoint));
-
-    return await res.json();
-  };
-
-  const { data: midpoint } = useQuery(['중간지점', participants], fetchMidpoint, { staleTime: Infinity });
-  const { data: stations } = useQuery(['지하철역', midpoint], fetchStations, {
-    enabled: !!midpoint,
-    staleTime: Infinity,
-  });
-  const closestStation = stations?.data?.[0];
+  const { stations, closestStation } = useMidpoint();
 
   useEffect(() => {
     if (closestStation) {
